requirements.txt
disnake
python-dotenv
bcrypt